Refactor D3 helpers into modular structure
