+++ src/d3-ext/constants.ts
@@ -1,0 +1,3 @@
+export const HANDLE_ICON_SIZE = 16;
+export const CONNECT_HANDLE_RADIUS = 4;
+export const SELECTION_STROKE_WIDTH = 1;
